Some ULONG -> SIZE_T / ULONG_PTR fixes
svn path=/branches/ros-amd64-bringup/; revision=36852

+++ reactos/ntoskrnl/mm/virtual.c
@@ -762,7 +762,7 @@ NtWriteVirtualMemory(IN HANDLE ProcessHandle,
     KPROCESSOR_MODE PreviousMode = ExGetPreviousMode();
     PEPROCESS Process;
     NTSTATUS Status = STATUS_SUCCESS;
-    ULONG BytesWritten = 0;
+    SIZE_T BytesWritten = 0;
     PAGED_CODE();    
 
     /* Check if we came from user mode */
